// lib/types/errors.ts
export enum RegistrationCheckInError {
  NOT_FOUND = "NOT_FOUND",
  ALREADY_CHECKED_IN = "ALREADY_CHECKED_IN",
  EVENT_ALREADY_ENDED = "EVENT_ALREADY_ENDED",
  EVENT_NOT_STARTED = "EVENT_NOT_STARTED",
  FORBIDDEN = "FORBIDDEN",
  NOT_PAID = "NOT_PAID",
  CODE_EXPIRED = "CODE_EXPIRED",
}

export enum AuthError {
  EMAIL_ADDRESS_INVALID = "EMAIL_ADDRESS_INVALID",
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  EMAIL_NOT_CONFIRMED = "EMAIL_NOT_CONFIRMED",
}